refactor(tui): extract viewport and layout sync component
